edit fixed, delete confirmation added

+++ app.js
@@ -6,6 +6,7 @@ let app = new Vue({
         albums: [],
         searchText: "",
         selectedYears: "Visi metai",
+        serverName: "http://localhost:3026",
         newAlbum: {
             artist: "",
             album: "",
@@ -13,9 +14,10 @@ let app = new Vue({
             genre: "",
             image: ""
         },
-        modalMode: {
-            edit: false,
-            new: false
+        modalEdit: {
+            visible: false,
+            albumId: "",
+            saved: false
         }
     },
     methods: {
@@ -24,27 +26,28 @@ let app = new Vue({
                 .then(function(response){
                     response.json()
                         .then(function(albums) {
-                            // Išsisaugom visus albumus
                             app.albums = albums;
- 
-                            // Spausdinam į HTML
-                            // renderAlbums(albums);
                         });
                 })
         },
         deleteAlbums: function (album) {
-            fetch(serverName + "/albums/" + album.id, {
-                method: "delete"
-            }).then(
-                function (response) {
-                    console.log("Albumas ištrintas")
-                }).catch(
-                function (error) {
-                    console.log("Albumas neištrintas", error);
-                });
+            let result = window.confirm("Ar tikrai norite ištrinti albumą?");
+            if(result) {
+                fetch(app.serverName + "/albums/" + album.id, {
+                    method: "delete",
+                }).then(
+                    function (response) {
+                        console.log("Albumas ištrintas")
+                    }).catch(
+                    function (error) {
+                        console.log("Albumas neištrintas", error);
+                    });
+    
+                let index = app.albums.indexOf(album);
+                app.albums.splice(index, 1);
 
-            let index = this.albums.indexOf(album);
-            this.albums.splice(index, 1);
+            }
+ 
         },
         getTags: function (text) {
             return text.split(", ");
@@ -53,52 +56,78 @@ let app = new Vue({
             this.newAlbum.image = event.target.files[0] ? event.target.files[0].name : "";
         },
         saveAlbum: function () {
-            fetch(serverName + "/albums", {
+            fetch(this.serverName + "/albums", {
                 method: "POST",
                 headers: { 'content-type': 'application/json' },
                 body: JSON.stringify(this.newAlbum)
             }).then(function (response) {
-                response.json()
+                response.json()                
                     .then(function (result) {
                         console.log("albumas išsaugotas");
                     });
             })
+            
             this.albums.push(this.newAlbum);
         },
         clearForm: function() {
-            this.modalMode.edit = false;
-            this.modalMode.new = false;
+            this.modalEdit.visible = false;
+            this.modalEdit.albumId = "";
+            this.modalEdit.saved = false;
 
-            app.newAlbum.album = "";
-            app.newAlbum.artist = "";
-            app.newAlbum.releaseDate = "";
-            app.newAlbum.genre = "";
+            this.newAlbum = {
+                artist: "",
+                album: "",
+                releaseDate: "",
+                genre: "",
+                image: ""
+            }
+          
         },
         editAlbum: function(album) {
-            this.modalMode.edit = true;
+            this.modalEdit.visible = true;
+            this.modalEdit.albumId = album.id;
 
-            this.newAlbum.artist = album.artist;
-            this.newAlbum.album = album.album;
-            this.newAlbum.releaseDate = album.releaseDate;
-            this.newAlbum.genre = album.genre;
-            this.newAlbum.image = album.image;
+            this.newAlbum = { 
+                artist: album.artist,
+                album: album.album,
+                releaseDate: album.releaseDate,
+                genre: album.genre,
+                image: album.image
+            }
         },
-        updateAlbumOnServer: function() {
-            // fetch(serverName + "/albums", {
-            //     method: "PUT",
-            //     headers: { 'content-type': 'application/json' },
-            //     body: JSON.stringify(this.newAlbum)
-            // }).then(function (response) {
-            //     response.json()
-            //         .then(function (result) {
-            //             console.log("albumas atnaujintas");
-            //         });
-            // })
-        // }
-            console.log("updated");
+        updateAlbum: function() {
+            // update on server
+            fetch(this.serverName + "/albums/" + this.modalEdit.albumId , {
+                method: "PUT",
+                headers: {
+                    'content-type': 'application/json'
+                },
+                body: JSON.stringify(this.newAlbum)
+            })
+            .then(function(response) {
+                response.json()
+                    .then(function(result) {
+                        console.log("albumas atnaujintas");
+                        app.modalEdit.saved = true;
+                    })
+            })
+            .catch(function(error) {
+                console.log("klaida: nepavyko atnaujinti albumo", error);
+            })
 
-    }
-},
+            // update UI with recent changes
+            app.albums.forEach(function(album){
+                if(app.modalEdit.albumId === album.id) {
+                    album.artist = app.newAlbum.artist;
+                    album.album = app.newAlbum.album;
+                    album.releaseDate = app.newAlbum.releaseDate;
+                    album.genre = app.newAlbum.genre;
+                    album.image = app.newAlbum.image
+                }
+            });
+        }   
+    },
+
     computed: {
         filteredAlbums: function () {
             let filteredByYear;
@@ -114,7 +143,7 @@ let app = new Vue({
                 let searchText = app.searchText.toLowerCase();
                 let resultArtist = item.artist.toLowerCase().indexOf(searchText) !== -1 ? true : false;
                 let resultAlbum = item.album.toLowerCase().indexOf(searchText) !== -1 ? true : false;
-                let resultGenre = item.genre.toLowerCase().indexOf(searchInput.value) !== -1 ? true : false;
+                let resultGenre = item.genre.toLowerCase().indexOf(searchText) !== -1 ? true : false;
                 if (resultArtist || resultAlbum || resultGenre) {
                     return true;
                 }
@@ -136,7 +165,7 @@ let app = new Vue({
 
 app.start();
 
-$('#newAlbum').on('hidden.bs.modal', function (e) {
-  app.clearForm();
-});
+// $('#newAlbum').on('hidden.bs.modal', function () {
+//   app.clearForm();
+// });
